requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pymilvus>=2.4.0
langchain-milvus
milvus-lite
langchain_google_genai
langchain-community
python-docx
pandas
requests
torch
numpy
pydub
python-multipart
DateTime
numpy
pandas
regex
yake
wordcloud
assemblyai
matplotlib
moviepy
Authlib
boto3
bs4
yt_dlp
docx2txt
fitz
pytest
langchain-nvidia-ai-endpoints
langchain-deepseek-official
openpyxl